Beaver Feeding Sign Locations recorded during the River Otter Beaver Trial.

A dataset (sf object) which includes the locations, survey times, survey season and impact level of each sign.

Usage

RivOtter_FeedSigns

Format

An sf object with 3683 rows and 4 variables:

SurveySeason

The winter season during which the sign was surveyed

RecordDate

The date that a sign was recorded

FeedCat

Character String describing the effort exerted by beaver on a given sign

geometry

The geometry data for the point location

Source

https://www.exeter.ac.uk/creww/research/beavertrial/