Estimate the number of territories in a catchment

Uses the kernel density maps to estimate the area of territories by locating areas of activity that contain foraging intensity above a given threshold, which locates potential territories - to confirm territories, confirmatory signs are required including dam and dwelling locations.

Usage

estimate_territories(
  forage_raster,
  confirm_signs,
  low_thresh = 0,
  upper_thresh = 0.95
)

Arguments

forage_raster

the foraging density raster generated from beavertools::forage_density()

confirm_signs

An sf object or an sf-readable file. See sf::st_drivers() for available drivers. containing 'confirmatory' field signs for beaver such as dams and lodges. This is required to assign a confidence classification for areas of activity either: 'Activity'. 'Possible' or 'Territory'.

low_thresh

numeric between 0 and 1. Used to set minimum value for activity. Lower values return more information i.e. area with cover but can result in the merging of territories as density increases. use lowest possible value.

upper_thresh

numeric between 0 and 1. Used to set the minimum value for core areas of activity. AS central place forages, beavers feed more near their lodge, this value idenitfies areas of activity which may indicate the centre of a territory. default is 0.95, may need adjusting depending on territory density.

Value

An 'sf' polygon object defining areas of beaver activity, with kernel denisty summary stats and the territory status.

Examples

# Here we filter the filter the built in 2019-2020 ROBT feeding sign data `RivOtter_FeedSigns`
# Then pipe this 'sf' object to forage_density.

ROBT_201920 <- RivOtter_FeedSigns %>%
dplyr::filter(SurveySeason == "2019 - 2020")%>%
  forage_density(., 'FeedCat')
#> No value supplied for "kd_extent" argument: default extent will be used
#> Warning: CRS object has comment, which is lost in output; in tests, see
#> https://cran.r-project.org/web/packages/sp/vignettes/CRS_warnings.html
#> 
#> calculating weighted kde
#> Warning: CRS object has comment, which is lost in output; in tests, see
#> https://cran.r-project.org/web/packages/sp/vignettes/CRS_warnings.html

# Now we load the ROBT `RivOtter_OtherSigns` dataset and filter to the same
# year as the forage density raster.

CS_201920 <- RivOtter_OtherSigns %>%
dplyr::filter(SurveySeason == "2019 - 2020")

# run territory classification
estimate_territories(ROBT_201920, confirm_signs = CS_201920)
#> Simple feature collection with 46 features and 8 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 305220 ymin: 83180 xmax: 323000 ymax: 114360
#> Projected CRS: OSGB 1936 / British National Grid
#> First 10 features:
#>     quant quantf id Upper_Thresh Confirm_signs terr_status      mean_fd
#> 1       0      0  1           No            No    Activity 1.380067e-12
#> 1.1     0      0  2           No            No    Activity 1.340421e-12
#> 1.2     0      0  3           No            No    Activity 1.034756e-12
#> 1.3     0      0  4           No            No    Activity 1.019890e-12
#> 1.4     0      0  5          Yes           Yes   Territory 2.339032e-07
#> 1.5     0      0  6          Yes           Yes   Territory 1.191316e-07
#> 1.6     0      0  7           No            No    Activity 1.427402e-12
#> 1.7     0      0  8           No            No    Activity 1.004674e-12
#> 1.8     0      0  9           No            No    Activity 1.018443e-12
#> 1.9     0      0 10          Yes            No    Possible 4.275207e-08
#>           sum_fd                       geometry
#> 1   7.590369e-11 POLYGON ((307500 83200, 307...
#> 1.1 7.372312e-11 POLYGON ((307500 83560, 307...
#> 1.2 9.312801e-12 POLYGON ((307540 83800, 307...
#> 1.3 9.179009e-12 POLYGON ((307600 83940, 307...
#> 1.4 1.022859e-03 POLYGON ((307920 84060, 307...
#> 1.5 1.934697e-04 POLYGON ((306020 84800, 306...
#> 1.6 7.422492e-11 POLYGON ((308720 86460, 308...
#> 1.7 3.014022e-12 POLYGON ((308640 88020, 308...
#> 1.8 3.055329e-12 POLYGON ((308520 88180, 308...
#> 1.9 1.285127e-04 POLYGON ((308840 88660, 308...